+++ db/schema.rb
@@ -33,7 +33,7 @@ ActiveRecord::Schema.define(version: 20160901002241) do
   create_table "reviews", force: :cascade do |t|
     t.integer  "business_id", null: false
     t.integer  "user_id",     null: false
-    t.integer  "rating",      null: false
+    t.float    "rating",      null: false
     t.string   "content",     null: false
     t.datetime "created_at",  null: false
     t.datetime "updated_at",  null: false
